+++ backend-lib/src/main/java/cash/z/ecc/android/sdk/internal/jni/RustBackend.kt
@@ -90,17 +90,43 @@ class RustBackend private constructor(
     }
 
     override suspend fun createAccount(
+        accountName: String,
+        keySource: String?,
         seed: ByteArray,
         treeState: ByteArray,
-        recoverUntil: Long?
+        recoverUntil: Long?,
     ): JniUnifiedSpendingKey {
         return withContext(SdkDispatchers.DATABASE_IO) {
             createAccount(
-                dataDbFile.absolutePath,
-                seed,
-                treeState,
-                recoverUntil ?: -1,
-                networkId = networkId
+                dbDataPath = dataDbFile.absolutePath,
+                networkId = networkId,
+                accountName = accountName,
+                keySource = keySource,
+                seed = seed,
+                treeState = treeState,
+                recoverUntil = recoverUntil ?: -1,
+            )
+        }
+    }
+
+    override suspend fun importAccountUfvk(
+        accountName: String,
+        keySource: String?,
+        ufvk: String,
+        treeState: ByteArray,
+        recoverUntil: Long?,
+        purpose: Int
+    ): JniAccount {
+        return withContext(SdkDispatchers.DATABASE_IO) {
+            importAccountUfvk(
+                dbDataPath = dataDbFile.absolutePath,
+                networkId = networkId,
+                accountName = accountName,
+                keySource = keySource,
+                ufvk = ufvk,
+                treeState = treeState,
+                recoverUntil = recoverUntil ?: -1,
+                purpose = purpose,
             )
         }
     }
@@ -114,11 +140,11 @@ class RustBackend private constructor(
             )
         }
 
-    override suspend fun getCurrentAddress(accountIndex: Int) =
+    override suspend fun getCurrentAddress(accountUuid: ByteArray) =
         withContext(SdkDispatchers.DATABASE_IO) {
             getCurrentAddress(
                 dataDbFile.absolutePath,
-                accountIndex,
+                accountUuid,
                 networkId = networkId
             )
         }
@@ -127,11 +153,11 @@ class RustBackend private constructor(
 
     override fun getSaplingReceiver(ua: String) = getSaplingReceiverForUnifiedAddress(ua)
 
-    override suspend fun listTransparentReceivers(accountIndex: Int): List<String> {
+    override suspend fun listTransparentReceivers(accountUuid: ByteArray): List<String> {
         return withContext(SdkDispatchers.DATABASE_IO) {
             listTransparentReceivers(
                 dbDataPath = dataDbFile.absolutePath,
-                accountIndex = accountIndex,
+                accountUuid = accountUuid,
                 networkId = networkId
             ).asList()
         }
@@ -318,14 +344,14 @@ class RustBackend private constructor(
     }
 
     override suspend fun proposeTransferFromUri(
-        accountIndex: Int,
+        accountUuid: ByteArray,
         uri: String
     ): ProposalUnsafe =
         withContext(SdkDispatchers.DATABASE_IO) {
             ProposalUnsafe.parse(
                 proposeTransferFromUri(
                     dataDbFile.absolutePath,
-                    accountIndex,
+                    accountUuid,
                     uri,
                     networkId = networkId,
                 )
@@ -333,7 +359,7 @@ class RustBackend private constructor(
         }
 
     override suspend fun proposeTransfer(
-        accountIndex: Int,
+        accountUuid: ByteArray,
         to: String,
         value: Long,
         memo: ByteArray?
@@ -342,7 +368,7 @@ class RustBackend private constructor(
             ProposalUnsafe.parse(
                 proposeTransfer(
                     dataDbFile.absolutePath,
-                    accountIndex,
+                    accountUuid,
                     to,
                     value,
                     memo,
@@ -352,7 +378,7 @@ class RustBackend private constructor(
         }
 
     override suspend fun proposeShielding(
-        accountIndex: Int,
+        accountUuid: ByteArray,
         shieldingThreshold: Long,
         memo: ByteArray?,
         transparentReceiver: String?
@@ -360,7 +386,7 @@ class RustBackend private constructor(
         return withContext(SdkDispatchers.DATABASE_IO) {
             proposeShielding(
                 dataDbFile.absolutePath,
-                accountIndex,
+                accountUuid,
                 shieldingThreshold,
                 memo,
                 transparentReceiver,
@@ -492,13 +518,29 @@ class RustBackend private constructor(
         ): Array<JniAccount>
 
         @JvmStatic
+        @Suppress("LongParameterList")
         private external fun createAccount(
             dbDataPath: String,
+            networkId: Int,
+            accountName: String,
+            keySource: String?,
             seed: ByteArray,
             treeState: ByteArray,
             recoverUntil: Long,
-            networkId: Int
         ): JniUnifiedSpendingKey
+
+        @JvmStatic
+        @Suppress("LongParameterList")
+        private external fun importAccountUfvk(
+            dbDataPath: String,
+            networkId: Int,
+            accountName: String,
+            keySource: String?,
+            ufvk: String,
+            treeState: ByteArray,
+            recoverUntil: Long,
+            purpose: Int,
+        ): JniAccount
 
         @JvmStatic
         private external fun isSeedRelevantToAnyDerivedAccounts(
@@ -510,7 +552,7 @@ class RustBackend private constructor(
         @JvmStatic
         private external fun getCurrentAddress(
             dbDataPath: String,
-            accountIndex: Int,
+            accountUuid: ByteArray,
             networkId: Int
         ): String
 
@@ -523,7 +565,7 @@ class RustBackend private constructor(
         @JvmStatic
         private external fun listTransparentReceivers(
             dbDataPath: String,
-            accountIndex: Int,
+            accountUuid: ByteArray,
             networkId: Int
         ): Array<String>
 
@@ -671,7 +713,7 @@ class RustBackend private constructor(
         @JvmStatic
         private external fun proposeTransferFromUri(
             dbDataPath: String,
-            accountIndex: Int,
+            accountUuid: ByteArray,
             uri: String,
             networkId: Int,
         ): ByteArray
@@ -680,7 +722,7 @@ class RustBackend private constructor(
         @Suppress("LongParameterList")
         private external fun proposeTransfer(
             dbDataPath: String,
-            accountIndex: Int,
+            accountUuid: ByteArray,
             to: String,
             value: Long,
             memo: ByteArray?,
@@ -691,7 +733,7 @@ class RustBackend private constructor(
         @Suppress("LongParameterList")
         private external fun proposeShielding(
             dbDataPath: String,
-            accountIndex: Int,
+            accountUuid: ByteArray,
             shieldingThreshold: Long,
             memo: ByteArray?,
             transparentReceiver: String?,