+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/ui/screen/home/viewmodel/WalletViewModel.kt
@@ -100,16 +100,18 @@ class WalletViewModel(application: Application) : AndroidViewModel(application) 
         secretState
             .filterIsInstance<SecretState.Ready>()
             .map { it.persistableWallet }
-            .map {
+            .map { secretState ->
                 val bip39Seed =
                     withContext(Dispatchers.IO) {
-                        Mnemonics.MnemonicCode(it.seedPhrase.joinToString()).toSeed()
+                        Mnemonics.MnemonicCode(secretState.seedPhrase.joinToString()).toSeed()
                     }
-                DerivationTool.getInstance().deriveUnifiedSpendingKey(
-                    seed = bip39Seed,
-                    network = it.network,
-                    account = getCurrentAccount()
-                )
+                getCurrentAccount().hdAccountIndex?.let { accountIndex ->
+                    DerivationTool.getInstance().deriveUnifiedSpendingKey(
+                        seed = bip39Seed,
+                        network = secretState.network,
+                        accountIndex = accountIndex.toInt()
+                    )
+                }
             }.stateIn(
                 viewModelScope,
                 SharingStarted.WhileSubscribed(ANDROID_STATE_FLOW_TIMEOUT),