+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/util/TestWallet.kt
@@ -6,6 +6,7 @@ import cash.z.ecc.android.bip39.toSeed
 import cash.z.ecc.android.sdk.SdkSynchronizer
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.WalletInitMode
+import cash.z.ecc.android.sdk.fixture.AccountCreateSetupFixture
 import cash.z.ecc.android.sdk.fixture.AccountFixture
 import cash.z.ecc.android.sdk.fixture.LightWalletEndpointFixture
 import cash.z.ecc.android.sdk.internal.Twig
@@ -64,19 +65,19 @@ class TestWallet(
             RustDerivationTool.new().deriveUnifiedSpendingKey(
                 seed = seed,
                 network = network,
-                accountIndex = AccountFixture.ZIP_32_ACCOUNT_INDEX
+                accountIndex = AccountFixture.ZIP_32_ACCOUNT_INDEX.toInt()
             )
         }
     val synchronizer: SdkSynchronizer =
         Synchronizer.newBlocking(
-            context,
-            network,
-            alias,
+            alias = alias,
+            birthday = startHeight,
+            context = context,
             lightWalletEndpoint = endpoint,
-            seed = seed,
-            startHeight,
+            setup = AccountCreateSetupFixture.new(),
             // Using existing wallet init mode as simplification for the test
-            walletInitMode = WalletInitMode.ExistingWallet
+            walletInitMode = WalletInitMode.ExistingWallet,
+            zcashNetwork = network,
         ) as SdkSynchronizer
 
     val available get() = synchronizer.walletBalances.value?.get(account)?.sapling?.available