+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Synchronizer.kt
@@ -17,6 +17,9 @@ import cash.z.ecc.android.sdk.internal.exchange.UsdExchangeRateFetcher
 import cash.z.ecc.android.sdk.internal.model.ext.toBlockHeight
 import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.AccountBalance
+import cash.z.ecc.android.sdk.model.AccountCreateSetup
+import cash.z.ecc.android.sdk.model.AccountImportSetup
+import cash.z.ecc.android.sdk.model.AccountPurpose
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.FastestServersResult
 import cash.z.ecc.android.sdk.model.ObserveFiatCurrencyResult
@@ -132,14 +135,28 @@ interface Synchronizer {
     val accountsFlow: Flow<List<Account>?>
 
     /**
-     * Measure connection quality and speed of given [servers].
+     * Tells the wallet to track an account using a unified full viewing key.
      *
-     * @return a [Flow] of fastest servers which updates it's state during measurement stages
+     * Returns details about the imported account, including the unique account identifier for
+     * the newly-created wallet database entry. Unlike the other account creation APIs, no spending key is returned
+     * because the wallet has no information about how the UFVK was derived.
+     *
+     * @param purpose Metadata describing whether or not data required for spending should be tracked by the wallet
+     * @param recoverUntil An optional height at which the wallet should exit "recovery mode"
+     * @param setup The account's setup information. See [AccountImportSetup] for more.
+     * @param treeState The tree state corresponding to the last block prior to the wallet's birthday height
+     *
+     * @return Account containing details about the imported account, including the unique account identifier for the
+     * newly-created wallet database entry
+     *
+     * @throws [InitializeException.ImportAccountException] in case of the operation failure
      */
-    suspend fun getFastestServers(
-        context: Context,
-        servers: List<LightWalletEndpoint>
-    ): Flow<FastestServersResult>
+    suspend fun importAccountByUfvk(
+        purpose: AccountPurpose,
+        recoverUntil: Long?,
+        setup: AccountImportSetup,
+        treeState: ByteArray,
+    ): Account
 
     /**
      * Adds the next available account-level spend authority, given the current set of
@@ -160,9 +177,9 @@ interface Synchronizer {
      * have been received by the currently-available account (in order to enable
      * automated account recovery).
      *
-     * @param seed the wallet's seed phrase.
-     * @param treeState
-     * @param recoverUntil
+     * @param recoverUntil An optional height at which the wallet should exit "recovery mode"
+     * @param setup The wallet's setup information. See [AccountCreateSetup] for more.
+     * @param treeState The tree state corresponding to the last block prior to the wallet's birthday height
      *
      * @return the newly created ZIP 316 account identifier, along with the binary
      * encoding of the `UnifiedSpendingKey` for the newly created account.
@@ -172,11 +189,21 @@ interface Synchronizer {
     @Suppress("standard:no-consecutive-comments")
     /* Not ready to be a public API; internal for testing only
     suspend fun createAccount(
-        seed: ByteArray,
+        setup: AccountCreateSetup,
         treeState: TreeState,
         recoverUntil: BlockHeight?
     ): UnifiedSpendingKey
      */
+
+    /**
+     * Measure connection quality and speed of given [servers].
+     *
+     * @return a [Flow] of fastest servers which updates it's state during measurement stages
+     */
+    suspend fun getFastestServers(
+        context: Context,
+        servers: List<LightWalletEndpoint>
+    ): Flow<FastestServersResult>
 
     /**
      * Gets the current unified address for the given account.
@@ -559,13 +586,13 @@ interface Synchronizer {
          * client wishes to change the server endpoint, the active synchronizer will need to be stopped and a new
          * instance created with a new value.
          *
-         * @param seed the wallet's seed phrase. This is required the first time a new wallet is set up. For
-         * subsequent calls, seed is only needed if [InitializerException.SeedRequired] is thrown.
-         *
          * @param birthday Block height representing the "birthday" of the wallet.  When creating a new wallet, see
          * [BlockHeight.ofLatestCheckpoint].  When restoring an existing wallet, use block height that was first used
          * to create the wallet.  If that value is unknown, null is acceptable but will result in longer
          * sync times.  After sync completes, the birthday can be determined from [Synchronizer.latestBirthdayHeight].
+         *
+         * @param setup An optional Account setup data that holds seed and other account related information.
+         * See [AccountCreateSetup] for more.
          *
          * @param walletInitMode a required parameter with one of [WalletInitMode] values. Use
          * [WalletInitMode.NewWallet] when starting synchronizer for a newly created wallet. Or use
@@ -585,13 +612,13 @@ interface Synchronizer {
          */
         @Suppress("LongParameterList", "LongMethod")
         suspend fun new(
-            context: Context,
-            zcashNetwork: ZcashNetwork,
             alias: String = ZcashSdk.DEFAULT_ALIAS,
-            lightWalletEndpoint: LightWalletEndpoint,
-            seed: ByteArray?,
             birthday: BlockHeight?,
-            walletInitMode: WalletInitMode
+            context: Context,
+            lightWalletEndpoint: LightWalletEndpoint,
+            setup: AccountCreateSetup?,
+            walletInitMode: WalletInitMode,
+            zcashNetwork: ZcashNetwork,
         ): CloseableSynchronizer {
             val applicationContext = context.applicationContext
 
@@ -652,9 +679,9 @@ interface Synchronizer {
                     rustBackend = backend,
                     databaseFile = coordinator.dataDbFile(zcashNetwork, alias),
                     checkpoint = loadedCheckpoint,
-                    seed = seed,
                     numberOfAccounts = Derivation.DEFAULT_NUMBER_OF_ACCOUNTS,
                     recoverUntil = chainTip,
+                    setup = setup,
                 )
 
             val encoder = DefaultSynchronizerFactory.defaultEncoder(backend, saplingParamTool, repository)
@@ -696,16 +723,24 @@ interface Synchronizer {
         @JvmStatic
         @Suppress("LongParameterList")
         fun newBlocking(
-            context: Context,
-            zcashNetwork: ZcashNetwork,
             alias: String = ZcashSdk.DEFAULT_ALIAS,
-            lightWalletEndpoint: LightWalletEndpoint,
-            seed: ByteArray?,
             birthday: BlockHeight?,
-            walletInitMode: WalletInitMode
+            context: Context,
+            lightWalletEndpoint: LightWalletEndpoint,
+            setup: AccountCreateSetup?,
+            walletInitMode: WalletInitMode,
+            zcashNetwork: ZcashNetwork,
         ): CloseableSynchronizer =
             runBlocking {
-                new(context, zcashNetwork, alias, lightWalletEndpoint, seed, birthday, walletInitMode)
+                new(
+                    alias = alias,
+                    birthday = birthday,
+                    context = context,
+                    lightWalletEndpoint = lightWalletEndpoint,
+                    setup = setup,
+                    walletInitMode = walletInitMode,
+                    zcashNetwork = zcashNetwork,
+                )
             }
 
         /**